Split PersistWorkflowInstanceDataCommand and PersistWorkflowInstanceDataEvent into their own handler and unit test
